requirements.txt
Flask==2.3.2
numpy==1.26.0
tensorflow==2.13.0
matplotlib==3.8.0
Pillow==10.0.1
Werkzeug==2.3.7
opencv-python==4.8.1.78
scikit-learn==1.3.0
tqdm==4.66.1